+++ src/App.js
@@ -1,5 +1,5 @@
 import React, { Component } from 'react';
-import Projects from './Projects/projects';
+import Projects from './Components/Projects/projects';
 import { Route } from 'react-router-dom';
 import './App.css';
 import Palette from './Components/Palette/palette'
@@ -17,8 +17,9 @@ class App extends Component {
       <div className="App">
       {/* <Route exact path='/palette' render={() => 
       <Palette key={Date.now()} />}/> */}
-      <Palette />
-      <Route exact path='/projects' render={() => <Projects />}/>
+      <Route exact path='/' render={() => <Projects />}/>
+      <Route exact path='/palettes/:id' render={() => <Palette />} />
+
       </div>
     );
   }
